+++ views/default/panels/user/summary.php
@@ -3,10 +3,10 @@
 ?>
 <div class="yii-debug-toolbar__block">
     <a href="<?= $panel->getUrl() ?>">
-        <?php if(Yii::$app->user->isGuest): ?>
+        <?php if (Yii::$app->user->isGuest): ?>
             <span class="yii-debug-toolbar__label">Guest</span>
         <?php else: ?>
-            <span class="yii-debug-toolbar__label yii-debug-toolbar__label_info">User <?= Yii::$app->user->id?></span>
-        <?php endif;?>
+            User <span class="yii-debug-toolbar__label yii-debug-toolbar__label_info"><?= Yii::$app->user->id ?></span>
+        <?php endif; ?>
     </a>
 </div>